requirements.txt
django
djangorestframework
coverage
drf-yasg
setuptools
pytest 
pytest-django
pytest-cov
python-dotenv
